renamed the vector type

+++ coding_vectorized/example.cpp
@@ -16,7 +16,7 @@
 
 #include <ap_int.h>
 
-typedef float float8 __attribute__((vector_size(64)));
+typedef float float16 __attribute__((vector_size(64)));
 
 template<int N, typename T>
 void load(T (&out)[N], const T *in, int o) {
@@ -48,15 +48,15 @@ void compute(T (&res)[N], const S (&lhs)[N], const S (&rhs)[N]) {
 #define N 32
 
 extern "C"
-void example(float8* res, const float8 *lhs, const float8 *rhs, int n) {
+void example(float16* res, const float16 *lhs, const float16 *rhs, int n) {
 #pragma HLS INTERFACE m_axi port=lhs offset=slave bundle=gmem0 depth=1000 // wordsize=64
 #pragma HLS INTERFACE m_axi port=rhs offset=slave bundle=gmem1 depth=1000 // wordsize=64
 #pragma HLS INTERFACE m_axi port=res offset=slave bundle=gmem0 depth=1000 // wordsize=64
 
   for(int i = 0; i < n; ++i) {
-    float8 lhs_buf[N];
-    float8 rhs_buf[N];
-    float8 res_buf[N];
+    float16 lhs_buf[N];
+    float16 rhs_buf[N];
+    float16 res_buf[N];
 
 #pragma HLS DATAFLOW
     load(lhs_buf, lhs, i);
